Gatekeeper-POC/Cohort-Endpoint/requirements.txt
django==2.2.13
#django==3.0.2
djangorestframework==3.9.1
#django-rest-auth==0.9.3
#django-allauth==0.39.1

numpy~=1.19.2
pandas~=1.1.2
six==1.9.0